Use s_* methods to implement {v_}compute()
